+++ packages/components/calendar/src/calendar.tsx
@@ -1,4 +1,4 @@
-import type {DateValue} from "@internationalized/date";
+import type {CalendarDate, DateValue} from "@internationalized/date";
 import type {ForwardedRef, ReactElement} from "react";
 
 import {forwardRef} from "@nextui-org/system";
@@ -7,7 +7,10 @@ import {UseCalendarProps, useCalendar} from "./use-calendar";
 import {CalendarProvider} from "./calendar-context";
 import {CalendarBase} from "./calendar-base";
 
-interface Props<T extends DateValue> extends Omit<UseCalendarProps<T>, "isHeaderWrapperExpanded"> {}
+interface Props<T extends DateValue>
+  extends Omit<UseCalendarProps<T>, "isHeaderWrapperExpanded" | "children"> {
+  children?: ((date: CalendarDate) => React.ReactNode) | React.ReactNode;
+}
 
 export type CalendarProps<T extends DateValue = DateValue> = Props<T>;
 
@@ -15,7 +18,12 @@ const Calendar = forwardRef(function Calendar<T extends DateValue>(
   props: CalendarProps<T>,
   ref: ForwardedRef<HTMLDivElement>,
 ) {
-  const {context, getBaseCalendarProps} = useCalendar<T>({...props, ref});
+  const {children, ...otherProps} = props;
+  const {context, getBaseCalendarProps} = useCalendar<T>({
+    ...otherProps,
+    ref,
+    cellContent: children,
+  });
 
   return (
     <CalendarProvider value={context}>