+++ packages/components/calendar/src/index.ts
@@ -1,5 +1,8 @@
 import Calendar from "./calendar";
 import RangeCalendar from "./range-calendar";
+import CalendarCellContent from "./calendar-cell-content";
+import CalendarCellHeader from "./calendar-cell-header";
+import CalendarCellBody from "./calendar-cell-body";
 
 // export types
 export type {CalendarProps} from "./calendar";
@@ -16,4 +19,4 @@ export {useRangeCalendar} from "./use-range-calendar";
 export {CalendarProvider, useCalendarContext} from "./calendar-context";
 
 // export component
-export {Calendar, RangeCalendar};
+export {Calendar, RangeCalendar, CalendarCellContent, CalendarCellHeader, CalendarCellBody};