+++ packages/components/calendar/src/range-calendar.tsx
@@ -1,4 +1,4 @@
-import type {DateValue} from "@internationalized/date";
+import type {CalendarDate, DateValue} from "@internationalized/date";
 import type {ForwardedRef, ReactElement} from "react";
 
 import {forwardRef} from "@nextui-org/system";
@@ -10,8 +10,10 @@ import {CalendarBase} from "./calendar-base";
 interface Props<T extends DateValue>
   extends Omit<
     UseRangeCalendarProps<T>,
-    "isHeaderExpanded" | "onHeaderExpandedChange" | "isHeaderWrapperExpanded"
-  > {}
+    "isHeaderExpanded" | "onHeaderExpandedChange" | "isHeaderWrapperExpanded" | "children"
+  > {
+  children?: ((date: CalendarDate) => React.ReactNode) | React.ReactNode;
+}
 
 export type RangeCalendarProps<T extends DateValue = DateValue> = Props<T>;
 
@@ -19,7 +21,12 @@ const RangeCalendar = forwardRef(function RangeCalendar<T extends DateValue>(
   props: RangeCalendarProps<T>,
   ref: ForwardedRef<HTMLDivElement>,
 ) {
-  const {context, getBaseCalendarProps} = useRangeCalendar<T>({...props, ref});
+  const {children, ...otherProps} = props;
+  const {context, getBaseCalendarProps} = useRangeCalendar<T>({
+    ...otherProps,
+    ref,
+    cellContent: children,
+  });
 
   return (
     <CalendarProvider value={context}>