+++ packages/components/date-picker/CHANGELOG.md
@@ -51,6 +51,7 @@
 
 ### Patch Changes
 
+
 - [#4255](https://github.com/nextui-org/nextui/pull/4255) [`6a94a12`](https://github.com/nextui-org/nextui/commit/6a94a125d4836b0a18d9cd2cb521c85a6bfa9050) Thanks [@wingkwong](https://github.com/wingkwong)! - fixed incorrect peerDependencies for theme and system package (#4254)
 
 - Updated dependencies [[`6a94a12`](https://github.com/nextui-org/nextui/commit/6a94a125d4836b0a18d9cd2cb521c85a6bfa9050)]:
@@ -60,6 +61,7 @@
   - @nextui-org/button@2.2.3
   - @nextui-org/form@2.1.3
   - @nextui-org/aria-utils@2.2.3
+
 
 ## 2.3.2